+++ examples/resources/elasticstack_elasticsearch_data_stream_lifecycle/import.sh
@@ -1,0 +1,2 @@
+terraform import elasticstack_elasticsearch_data_stream_lifecycle.my_data_stream_lifecycle <cluster_uuid>/<data_stream_name>
+